+++ packages/core/src/components/TextField/Styled/Prefix.styled.tsx
@@ -1,6 +1,6 @@
 import { InjectClassName, styled } from '@medly-components/utils';
 
-export const Prefix = styled(InjectClassName)`
+export const Prefix = styled(InjectClassName)<{ size?: 'S' | 'M' }>`
     pointer-events: none;
-    margin-right: 1.6rem;
+    margin-right: ${({ size }) => (size === 'S' ? '1.2rem' : '1.6rem')};
 `;