+++ packages/theme/CHANGELOG.md
@@ -3,6 +3,28 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [1.13.0](https://github.com/medly/medly-components/compare/@medly-components/theme@1.12.0...@medly-components/theme@1.13.0) (2020-08-29)
+
+
+### Features
+
+* move searchbox text variant to theme ([245a0ea](https://github.com/medly/medly-components/commit/245a0eaba62416ba02c410a4388136924e94993b))
+
+
+
+
+
+# [1.12.0](https://github.com/medly/medly-components/compare/@medly-components/theme@1.11.0...@medly-components/theme@1.12.0) (2020-08-29)
+
+
+### Features
+
+* add small styles for textfield, singleSelect & multiselect component ([d4658fc](https://github.com/medly/medly-components/commit/d4658fc39b42a01d54aee6002db2733642a0eb77))
+
+
+
+
+
 # [1.11.0](https://github.com/medly/medly-components/compare/@medly-components/theme@1.10.0...@medly-components/theme@1.11.0) (2020-08-27)